+++ src/lib/barrels.ts
@@ -1,8 +1,25 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { PipesModule } from './pipes/pipes.module';
 import { StBreadcrumbsModule } from './st-breadcrumbs/st-breadcrumbs.module';
 import { StButtonModule } from './st-button/st-button.module';
-import { StDropdownModule } from './st-dropdown/st-dropdown.module';
+import { StCheckboxModule } from './st-checkbox/st-checkbox.module';
+import { StComboboxModule } from './st-combobox/st-combobox.module';
 import { StDropdownMenuModule } from './st-dropdown-menu/st-dropdown-menu.module';
+import { StDropdownModule } from './st-dropdown/st-dropdown.module';
 import { StFooterModule } from './st-footer/st-footer.module';
 import { StFormModule } from './st-form/st-form.module';
 import { StFormLabelModule } from './utils/egeo-form/st-form-label/st-form-label.module';
@@ -14,10 +31,12 @@ import { StInputModule } from './st-input/st-input.module';
 import { StModalModule } from './st-modal/st-modal.module';
 import { StPageTitleModule } from './st-page-title/st-page-title.module';
 import { StPaginationModule } from './st-pagination/st-pagination.module';
-import { StRadioModule } from './st-radio/st-radio.module';
+import { StPopModule } from './st-pop/st-pop.module';
 import { StRadioMenuModule } from './st-radio-menu/st-radio-menu.module';
+import { StRadioModule } from './st-radio/st-radio.module';
 import { StSearchModule } from './st-search/st-search.module';
 import { StSpinnerModule } from './st-spinner/st-spinner.module';
+import { StSwitchModule } from './st-switch/st-switch.module';
 import { StTabBoxModule } from './st-tab-box/st-tab-box.module';
 import { StTableModule } from './st-table/st-table.module';
 import { StTextareaModule } from './st-textarea/st-textarea.module';
@@ -25,16 +44,15 @@ import { StToggleButtonsModule } from './st-toggle-buttons/st-toggle-buttons.mod
 import { StTooltipModule } from './st-tooltip/st-tooltip.module';
 import { StTwoListSelectionModule } from './st-two-list-selection/st-two-list-selection.module';
 import { StVerticalTabsModule } from './st-vertical-tabs/st-vertical-tabs.module';
-import { StSwitchModule } from './st-switch/st-switch.module';
-import { StCheckboxModule } from './st-checkbox/st-checkbox.module';
-import { StPopModule } from './st-pop/st-pop.module';
 
 export const DECLARATIONS: any[] = [
    PipesModule,
    StBreadcrumbsModule,
    StButtonModule,
-   StDropdownModule,
+   StCheckboxModule,
+   StComboboxModule,
    StDropdownMenuModule,
+   StDropdownModule,
    StFooterModule,
    StFormModule,
    StFormLabelModule,
@@ -43,21 +61,20 @@ export const DECLARATIONS: any[] = [
    StInfoBoxModule,
    StInfoCardModule,
    StInputModule,
+   StModalModule,
    StPageTitleModule,
+   StPaginationModule,
+   StPopModule,
    StRadioMenuModule,
+   StRadioModule,
    StSearchModule,
    StSpinnerModule,
+   StSwitchModule,
    StTabBoxModule,
+   StTableModule,
+   StTextareaModule,
    StToggleButtonsModule,
    StTooltipModule,
    StTwoListSelectionModule,
-   StVerticalTabsModule,
-   StPaginationModule,
-   StModalModule,
-   StRadioModule,
-   StSwitchModule,
-   StTableModule,
-   StTextareaModule,
-   StCheckboxModule,
-   StPopModule
+   StVerticalTabsModule
 ];