+++ src/lib/public_api.ts
@@ -1,18 +1,47 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 /**************************** MAIN MODULES *********************************************/
 export { EgeoModule } from './egeo.module';
 
 /**************************** UTILS AND OTHERS *****************************************/
-export { StEgeo, StRequired, StDeprecated } from './decorators/require-decorators';
-export { TranslateableElement, EgeoResolverKeys, TranslateServiceType } from './utils/egeo-resolver/egeo-resolve-model';
+export {
+   StEgeo,
+   StRequired,
+   StDeprecated
+} from './decorators/require-decorators';
+export {
+   TranslateableElement,
+   EgeoResolverKeys,
+   TranslateServiceType
+} from './utils/egeo-resolver/egeo-resolve-model';
 export { EgeoResolveService } from './utils/egeo-resolver/egeo-resolve.service';
 export { EgeoUtils } from './utils/egeo-utils';
 export { EventWindowManager } from './utils/event-window-manager';
 export { StRegEx } from './utils/st-regex';
-export { SelectOneDispaptcherListener, SelectOneDispatcher } from './utils/unique-dispatcher';
+export {
+   SelectOneDispaptcherListener,
+   SelectOneDispatcher
+} from './utils/unique-dispatcher';
 
 /**************************** PIPES ***************************************************/
 export { PipesModule } from './pipes/pipes.module';
 export { StFilterList } from './pipes/search-filter.pipe';
+
+/**************************** DIRECTIVES ***************************************************/
+export { StInputAdjustable } from './directives';
 
 /**************************** MODULES *************************************************/
 
@@ -22,12 +51,25 @@ export { StBreadcrumbsModule } from './st-breadcrumbs/st-breadcrumbs.module';
 // Button
 export { StButtonModule } from './st-button/st-button.module';
 
+// Checkbox
+export { StCheckboxModule } from './st-checkbox/st-checkbox.module';
+
+// Combobox
+export { StComboboxModule } from './st-combobox/st-combobox.module';
+
 // Dropdown
 export { StDropdownModule } from './st-dropdown/st-dropdown.module';
 
 // Dropdown menu
-export { StDropDownMenuGroup, StDropDownMenuGroupSchema, StDropDownMenuItem, StDropDownMenuItemSchema } from './st-dropdown-menu/st-dropdown-menu.interface';
-export { StDropdownMenuModule } from './st-dropdown-menu/st-dropdown-menu.module';
+export {
+   StDropDownMenuGroup,
+   StDropDownMenuGroupSchema,
+   StDropDownMenuItem,
+   StDropDownMenuItemSchema
+} from './st-dropdown-menu/st-dropdown-menu.interface';
+export {
+   StDropdownMenuModule
+} from './st-dropdown-menu/st-dropdown-menu.module';
 
 // Footers
 export { StFooterModule } from './st-footer/st-footer.module';
@@ -45,8 +87,13 @@ export {
 export { StHeaderModule } from './st-header/st-header.module';
 
 // Horizontal Tabs
-export { StHorizontalTabsModule } from './st-horizontal-tabs/st-horizontal-tabs.module';
-export { StHorizontalTab, StHorizontalTabSchema } from './st-horizontal-tabs/st-horizontal-tabs.model';
+export {
+   StHorizontalTabsModule
+} from './st-horizontal-tabs/st-horizontal-tabs.module';
+export {
+   StHorizontalTab,
+   StHorizontalTabSchema
+} from './st-horizontal-tabs/st-horizontal-tabs.model';
 
 // Info box
 export { StInfoBoxModule } from './st-info-box/st-info-box.module';
@@ -56,7 +103,10 @@ export { StInfoCardModule } from './st-info-card/st-info-card.module';
 
 // Input
 export { StInputModule } from './st-input/st-input.module';
-export { StInputError, StInputErrorSchema } from './st-input/st-input.error.model';
+export {
+   StInputError,
+   StInputErrorSchema
+} from './st-input/st-input.error.model';
 
 // Modal
 export {
@@ -74,7 +124,11 @@ export { StModalModule } from './st-modal/st-modal.module';
 export { StPageTitleModule } from './st-page-title/st-page-title.module';
 
 // Pagination
-export { Paginate, PaginateTexts, PaginateTextsSchema } from './st-pagination/st-pagination.interface';
+export {
+   Paginate,
+   PaginateTexts,
+   PaginateTextsSchema
+} from './st-pagination/st-pagination.interface';
 export { StPaginationPipe } from './st-pagination/st-pagination.pipe';
 export { StPaginationService } from './st-pagination/st-pagination.service';
 export { StPaginationModule } from './st-pagination/st-pagination.module';
@@ -85,7 +139,10 @@ export { RadioChange } from './st-radio/st-radio.change';
 
 // Radio menu
 export { StRadioMenuModule } from './st-radio-menu/st-radio-menu.module';
-export { StRadioMenuOption, StRadioMenuOptionSchema } from './st-radio-menu/st-radio-menu-option.interface';
+export {
+   StRadioMenuOption,
+   StRadioMenuOptionSchema
+} from './st-radio-menu/st-radio-menu-option.interface';
 
 // Search
 export { StSearchModule } from './st-search/st-search.module';
@@ -106,18 +163,28 @@ export { StTableHeader } from './st-table/shared/table-header.interface';
 export { Order, ORDER_TYPE } from './st-table/shared/order';
 
 // Textarea
-export { StTextareaError, StTextareaErrorSchema } from './st-textarea/st-textarea.error.model';
+export {
+   StTextareaError,
+   StTextareaErrorSchema
+} from './st-textarea/st-textarea.error.model';
 export { StTextareaModule } from './st-textarea/st-textarea.module';
 
 // Toogle buttons
-export { StToggleButton, StToggleButtonSchema } from './st-toggle-buttons/st-toggle-buttons.interface';
-export { StToggleButtonsModule } from './st-toggle-buttons/st-toggle-buttons.module';
+export {
+   StToggleButton,
+   StToggleButtonSchema
+} from './st-toggle-buttons/st-toggle-buttons.interface';
+export {
+   StToggleButtonsModule
+} from './st-toggle-buttons/st-toggle-buttons.module';
 
 // Tooltip
 export { StTooltipModule } from './st-tooltip/st-tooltip.module';
 
 // Two list selection
-export { StTwoListSelectionModule } from './st-two-list-selection/st-two-list-selection.module';
+export {
+   StTwoListSelectionModule
+} from './st-two-list-selection/st-two-list-selection.module';
 export {
    StTwoListSelectionConfig,
    StTwoListSelectionConfigSchema,
@@ -125,10 +192,13 @@ export {
    StTwoListSelectionAction,
    StTwoListSelectExtraLabelAction
 } from './st-two-list-selection/st-two-list-selection.model';
-export { StTwoListSelection } from './st-two-list-selection/st-two-list-selection';
+export {
+   StTwoListSelection
+} from './st-two-list-selection/st-two-list-selection';
 
 // Vertical tabs
-export { StVerticalTabsModule } from './st-vertical-tabs/st-vertical-tabs.module';
+export {
+   StVerticalTabsModule
+} from './st-vertical-tabs/st-vertical-tabs.module';
 
-export { StCheckboxModule } from './st-checkbox/st-checkbox.module';
 export { StPopModule } from './st-pop/st-pop.module';