+++ src/lib/st-checkbox/st-checkbox.component.ts
@@ -1,8 +1,24 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { Component, Input, Output, EventEmitter, forwardRef, ChangeDetectorRef, ChangeDetectionStrategy } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 
 export const CHECKBOX_CONTROL_ACCESSOR: any = {
    provide: NG_VALUE_ACCESSOR,
+   // tslint:disable-next-line:no-use-before-declare
    useExisting: forwardRef(() => StCheckboxComponent),
    multi: true
 };
@@ -60,7 +76,7 @@ export class StCheckboxComponent implements ControlValueAccessor {
       }
    }
 
-   handleChange($event): void {
+   handleChange($event: Event): void {
       $event.stopPropagation();
    }