+++ src/lib/st-radio-menu/st-radio-menu.component.spec.ts
@@ -1,21 +1,33 @@
-import {
-   TestBed
-} from '@angular/core/testing';
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { TestBed } from '@angular/core/testing';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 
-import {Http} from '@angular/http';
-import {
-   RouterTestingModule
-} from '@angular/router/testing';
-import {StRadioMenuOption} from './st-radio-menu-option.interface';
-import {StRadioMenuComponent} from './st-radio-menu.component';
-
+import { Http } from '@angular/http';
+import { RouterTestingModule } from '@angular/router/testing';
+import { StRadioMenuOption } from './st-radio-menu-option.interface';
+import { StRadioMenuComponent } from './st-radio-menu.component';
 
 describe('StRadioMenuComponent', () => {
    beforeEach(() => {
       TestBed.configureTestingModule({
          declarations: [StRadioMenuComponent],
          imports: [RouterTestingModule],
-         providers: []
+         providers: [],
+         schemas: [NO_ERRORS_SCHEMA]
       });
    });
 
@@ -24,8 +36,9 @@ describe('StRadioMenuComponent', () => {
    beforeEach(() => {
       stRadioMenuComponent = new StRadioMenuComponent();
       activeOption = {
-      label: 'active option name',
-      value: 'active option value'};
+         label: 'active option name',
+         value: 'active option value'
+      };
       stRadioMenuComponent.activeOption = activeOption;
    });
 
@@ -50,8 +63,9 @@ describe('StRadioMenuComponent', () => {
 
          stRadioMenuComponent.activateOption(activeOption);
 
-         expect(stRadioMenuComponent.changedOption.emit).toHaveBeenCalledWith(activeOption);
+         expect(stRadioMenuComponent.changedOption.emit).toHaveBeenCalledWith(
+            activeOption
+         );
       });
-
    });
 });