+++ src/lib/st-radio-menu/st-radio-menu.module.ts
@@ -1,12 +1,29 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { StRadioMenuComponent } from './st-radio-menu.component';
+import { StRadioModule } from '../st-radio/st-radio.module';
+import { SelectOneDispatcher } from '../utils/unique-dispatcher';
 
 @NgModule({
-   imports: [CommonModule, FormsModule, ReactiveFormsModule],
+   imports: [CommonModule, StRadioModule],
    declarations: [StRadioMenuComponent],
-   exports: [StRadioMenuComponent]
+   exports: [StRadioMenuComponent],
+   providers: [SelectOneDispatcher]
 })
-export class StRadioMenuModule { }
+export class StRadioMenuModule {}