+++ src/lib/st-radio/st-radio.component.ts
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import {
    Component,
    Input,
@@ -15,9 +30,9 @@ import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { SelectOneDispatcher } from '../utils/unique-dispatcher';
 import { RadioChange } from './st-radio.change';
 
-
 export const MD_RADIO_GROUP_CONTROL_VALUE_ACCESSOR: any = {
    provide: NG_VALUE_ACCESSOR,
+   // tslint:disable-next-line:no-use-before-declare
    useExisting: forwardRef(() => StRadioGroupComponent),
    multi: true
 };
@@ -80,6 +95,7 @@ export class StRadioGroupComponent implements OnInit, ControlValueAccessor {
       this._disabled = (value != null && value !== false) ? true : null;
    }
 
+   // tslint:disable-next-line:no-use-before-declare
    @ContentChildren(forwardRef(() => StRadioComponent))
    _radios: QueryList<StRadioComponent> = null;