+++ src/lib/st-two-list-selection/list-selection/list-selection.component.spec.ts
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { DebugElement, SimpleChange, SimpleChanges } from '@angular/core';
 import { async, ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { Http } from '@angular/http';
@@ -10,6 +25,7 @@ import {
 
 // Components
 import { ListItemComponent } from '../list-item/list-item.component';
+import { ListScrollComponent } from '../list-scroll/list-scroll.component';
 import { ListSelectionComponent } from './list-selection.component';
 
 // Order modules
@@ -43,7 +59,7 @@ describe('StTwoListSelectionComponent', () => {
    beforeEach(async(() => {
       TestBed.configureTestingModule({
          imports: [StSearchModule, VirtualScrollModule, StDropdownModule, StCheckboxModule],
-         declarations: [ListSelectionComponent, ListItemComponent]
+         declarations: [ListSelectionComponent, ListItemComponent, ListScrollComponent]
       })
          .compileComponents();  // compile template and css
    }));
@@ -61,11 +77,6 @@ describe('StTwoListSelectionComponent', () => {
          expect(comp.searchQaTag).toEqual(qaTag + '-search');
          expect(comp.listQaTag).toEqual(qaTag + '-list');
          expect(comp.hasOrder).toBeFalsy();
-         expect(comp.heightMode).toEqual(35);
-
-         comp.mode = 'compact';
-         fixture.detectChanges();
-         expect(comp.heightMode).toEqual(27);
       });
    });
 });