+++ src/lib/st-two-list-selection/st-two-list-selection.component.spec.ts
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { DebugElement, SimpleChange, SimpleChanges } from '@angular/core';
 import { async, ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { Http } from '@angular/http';
@@ -13,6 +28,7 @@ import { ListItemComponent } from './list-item/list-item.component';
 import { ListSelectionComponent } from './list-selection/list-selection.component';
 import { StTwoListSelectionViewComponent } from './st-two-list-selection.view.component';
 import { StTwoListSelectionComponent } from './st-two-list-selection.component';
+import { ListScrollComponent } from './list-scroll/list-scroll.component';
 
 // Order modules
 import { PipesModule } from '../pipes/pipes.module';
@@ -52,7 +68,7 @@ describe('StTwoListSelectionComponent', () => {
    beforeEach(async(() => {
       TestBed.configureTestingModule({
          imports: [StSearchModule, PipesModule, VirtualScrollModule, StButtonModule, StDropdownModule, StCheckboxModule],
-         declarations: [StTwoListSelectionComponent, StTwoListSelectionViewComponent, ListSelectionComponent, ListItemComponent]
+         declarations: [StTwoListSelectionComponent, StTwoListSelectionViewComponent, ListSelectionComponent, ListItemComponent, ListScrollComponent]
       })
          .compileComponents();  // compile template and css
    }));