+++ src/lib/st-two-list-selection/st-two-list-selection.module.ts
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2016 Stratio (http://stratio.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { VirtualScrollModule } from 'angular2-virtual-scroll';
@@ -5,6 +20,7 @@ import { VirtualScrollModule } from 'angular2-virtual-scroll';
 // Components
 import { ListItemComponent } from './list-item/list-item.component';
 import { ListSelectionComponent } from './list-selection/list-selection.component';
+import { ListScrollComponent } from './list-scroll/list-scroll.component';
 import { StTwoListSelectionComponent } from './st-two-list-selection.component';
 import { StTwoListSelectionViewComponent } from './st-two-list-selection.view.component';
 
@@ -17,7 +33,7 @@ import { StCheckboxModule } from '../st-checkbox/st-checkbox.module';
 
 @NgModule({
    imports: [CommonModule, StSearchModule, PipesModule, VirtualScrollModule, StButtonModule, StDropdownModule, StCheckboxModule],
-   declarations: [StTwoListSelectionViewComponent, StTwoListSelectionComponent, ListSelectionComponent, ListItemComponent],
+   declarations: [StTwoListSelectionViewComponent, StTwoListSelectionComponent, ListSelectionComponent, ListItemComponent, ListScrollComponent],
    exports: [StTwoListSelectionComponent, StTwoListSelectionViewComponent]
 })
 export class StTwoListSelectionModule { }