+++ tools/gulp/util/task_helpers.ts
@@ -12,7 +12,7 @@ const gulpIf = require('gulp-if');
 const gulpPostCss = require('gulp-postcss');
 const gulpRunSequence = require('run-sequence');
 const gulpSass = require('gulp-sass');
-const gulpSourcemaps = require('gulp-sourcemaps');
+const removeHtmlComments = require('gulp-remove-html-comments');
 
 // There are no type definitions available for these imports.
 const resolveBin = require('resolve-bin');
@@ -41,11 +41,9 @@ export function tsBuildTask(tsConfigPath: string): any {
 export function sassBuildTask(dest: string, root: string, minify: boolean = false): any {
    return () => {
       return gulp.src(_globify(root, '**/*.scss'))
-         .pipe(gulpSourcemaps.init({ loadMaps: true }))
          .pipe(gulpSass().on('error', gulpSass.logError))
          .pipe(gulpPostCss([autoprefixer()]))
          .pipe(gulpIf(minify, gulpCleanCss()))
-         .pipe(gulpSourcemaps.write('.'))
          .pipe(gulp.dest(dest));
    };
 }
@@ -117,9 +115,9 @@ export function execNodeTask(packageName: string, executable: string | string[],
 /** Copy files from a glob to a destination. */
 export function copyTask(srcGlobOrDir: string | string[], outRoot: string): any {
    if (typeof srcGlobOrDir === 'string') {
-      return () => gulp.src(_globify(srcGlobOrDir)).pipe(gulp.dest(outRoot));
+      return () => gulp.src(_globify(srcGlobOrDir)).pipe(removeHtmlComments()).pipe(gulp.dest(outRoot));
    } else {
-      return () => gulp.src(srcGlobOrDir.map(name => _globify(name))).pipe(gulp.dest(outRoot));
+      return () => gulp.src(srcGlobOrDir.map(name => _globify(name))).pipe(removeHtmlComments()).pipe(gulp.dest(outRoot));
    }
 }